feat: simplify badge routes and remove titles
- Change from /badges/[slug]/level-[level] to /badges/[slug]
- Show current level only for earned badges
- Remove page title and subtitle from profile badges tab
- Update sharing URLs to use simplified route structure
- Only earned badges (level > 0) are publicly accessible
- Clean up badge grid layout without headers

New simplified routes:
- /[identifier]/badges/[slug] - Public badge page showing current level
- No level parameter needed - much cleaner URLs

Benefits:
- Simpler URL structure
- Current level auto-display
- Better user experience
- Cleaner sharing URLs
